backend/pyproject.toml
[project]
name = "techbible-hack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "langchain-core>=1.2.23",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "mcp>=1.26.0",
    "pydantic-settings>=2.13.1",
    "sse-starlette>=3.3.4",
    "uvicorn>=0.42.0",
]
